Fixed for checkpoints

+++ src/etc/smartrouter/update.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
-rm /etc/smartrouter/dnsmasq_list.conf
-rm /etc/dnsmasq.d/dnsmasq_list.conf
 
-wget www.mitang.me/dnsmasq_list.conf -O /etc/smartrouter/dnsmasq_list.conf
-cp /etc/smartrouter/dnsmasq_list.conf /etc/dnsmasq.d/
+wget --spider --quiet --tries=1 --timeout=3 www.mitang.me
+if  [ "$?" == "0" ]
+then
+	wget www.mitang.me/dnsmasq_list.conf -O /etc/smartrouter/dnsmasq_list.conf	
+	cp /etc/smartrouter/dnsmasq_list.conf /etc/dnsmasq.d/
+	rm /etc/smartrouter/dnsmasq_list.conf
+fi
